requirements.txt
email-validator
beautifulsoup4
requests
numpy
pandas
scikit-learn
tensorflow
python-dotenv